% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsoap_svg.R
\name{get_circle_geom}
\alias{get_circle_geom}
\title{A function to retrieve and compose layer information for interactive
plotting.}
\usage{
get_circle_geom(name)
}
\description{
This function along with [geom_circle_interactive] is an addition to the
base functionalities of ggiraph package.
}
