% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsoap_layout.R
\name{gsoap_layout}
\alias{gsoap_layout}
\title{A function to create a layout for GSOAP plot}
\usage{
gsoap_layout(x, genes, pvalues, splitter = "/", distance = "jaccard",
  projection = "iso", scale.factor = 0.8, weighted = TRUE,
  packing = TRUE, clustering = TRUE, hc.method = "ward.D",
  no.clusters = NULL, max.clusters = 8, cluster.stat = "meta",
  isomap.k = 3, tsne.perplexity = 30, tsne.iterations = 1000,
  cca.epochs = 10, cca.alpha0 = 0.5)
}
\arguments{
\item{x}{a data frame with the results of gene set over-representation analysis.
Must have a rownames indicating names of the genes sets and at least two columns,
one of which contains query gene members; and another one that contains respective
p-values (raw or adjusted for multiple testing).}

\item{genes}{a character or integer, indicating name or index of the column containing genes members.}

\item{pvalues}{a character or integer, indicating name or index of the column containing p-values.}

\item{splitter}{a character to be used as a delimiter to parse the genes column.}

\item{distance}{a character indicating method used to calculate the distance/dissimilarity between instances.
Options include (but are not limited to) \emph{jaccard} (default), \emph{manhattan}, \emph{dice},
\emph{pearson}. For more details see \code{\link[philentropy]{distance}}.}

\item{projection}{a character indicating method used to project instances into 2-dimensional space based on their distance/dissimilarity..
Ooptions include \emph{iso} (isomap; default), \emph{mds} (multidimensional scaling), \emph{cca} (curvilinear component analysis), \emph{tsne} (t-distributed stochastic neighbor embedding),}

\item{scale.factor}{a positive real number to control dependence of the circle radius on the number of query gene members of the given gene set.}

\item{weighted}{a boolean indicating whether to use pathway \emph{importance}
(-log10(pvalue)) as a weight when closeness and clustering are calculated.}

\item{packing}{a boolean indicating whether to apply circle packing.}

\item{clustering}{a boolean indicating whether to apply clustering.}

\item{hc.method}{a character indicating method of hierarchical cluster to be used.
Options include: \emph{ward.D} (default), \emph{ward.D2}, \emph{single}, \emph{complete},
\emph{average}, \emph{mcquitty}, \emph{median} and \emph{centroid}.}

\item{no.clusters}{an integer indicating number of clusters, must be less than number of gene sets (rows).}

\item{max.clusters}{an integer indicating maximum number of clusters to consider, must be at least two.}

\item{cluster.stat}{an indicating statistic used to select optimal number of clusters.
Options are:
\itemize{
    \item \emph{meta} (default) is a combination of the methods listed below
    \item \emph{PBC} (point biserial correlation; default)
    \item \emph{HG} (Hubert's samma)
    \item \emph{HGSD} (Hubert’s samma - Somer's D)
    \item \emph{ASW} (average silhouette width)
    \item \emph{ASWw} (Average weighted silhouette width)
    \item \emph{CH} (Calinski-Harabasz index)
    \item \emph{R2} (R-squared)
    \item \emph{CHsq} (Calinski-Harabasz index using squared distances)
    \item \emph{R2sq} (R-squared using squared distances)
    \item \emph{HC} (Hubert’s C coefficient)
}}

\item{isomap.k}{an integer indicating number of k nearest neighbors of
the \emph{isomap} projection.}

\item{tsne.perplexity}{an integer indicating \emph{tSNE} perplexity.}

\item{tsne.iterations}{an integer indicating maximum number of \emph{tSNE} iterations to perform.}

\item{cca.epochs}{an integer indicating \emph{CCA} training length.}

\item{cca.alpha0}{a positive real number indicating \emph{CCA} initial step size.}
}
\value{
\code{layout} a data frame with x and y coordinates of
    the points representing the insttances, their size (radius) derived from
    the number of gene members; importance (-log10(p-value)), closeness,
    cluster membership and intracluster closeness.
}
\description{
Some description
}
\details{
More description
}
\examples{
data(pxgenes)

l = gsoap_layout(pxgenes, 'Members', 'p.value')

head(l$layout)
}
\author{
Tomas Tokar <tomastokar@gmail.com>
}
