% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsoap_plot.R
\name{gsoap_plot}
\alias{gsoap_plot}
\title{A function to create Gene Set Overrepresentation Analysis Plot}
\usage{
gsoap_plot(layout, as.color = NULL, as.alpha = NULL,
  which.labels = NULL, show.size.guide = TRUE,
  show.color.guide = TRUE, show.alpha.guide = TRUE,
  size.guide.loc = c(1, 0), size.guide.no = 4,
  default.color = "black", default.alpha = 0.5, range.alpha = c(0.1,
  0.9), viridis.option = "viridis", viridis.direction = 1,
  viridis.range = c(0, 1), label.alpha = 0.6, segment.alpha = 0.5,
  repel.xynudges = c(0.1, 0.1), repel.hvjust = c(0, 0),
  repel.direction = "both", base.fontsize = 8,
  size.guide.fontsize = base.fontsize - 2,
  label.fontsize = base.fontsize - 1, title = NULL, subtitle = NULL,
  xlabel = "proj. 1", ylabel = "proj. 2", xlimits = c(-0.1, 1.1),
  ylimits = c(-0.1, 1.1), void = FALSE)
}
\arguments{
\item{layout}{a data frame, a layout created by \code{\link{gsoap_layout}}.}

\item{as.color}{a character or integer, indicating name or index of the column with color aesthetic.}

\item{as.alpha}{a character or integer, indicating name or index of the column with alpha aesthetic.}

\item{which.labels}{a character or integer vector, indicating gene sets (rows)
whose names (rownames) should be show in the plot}

\item{show.size.guide}{a boolean indicating whether to show size guide.}

\item{show.color.guide}{a boolean indicating whether to show color guide.}

\item{size.guide.loc}{a numeric vector of length 2 with the size guide x and y coordinates.}

\item{size.guide.no}{an integer indicating number of circles of the size guide.}

\item{default.color}{a character indicating color to be used if no color aesthetics is used.}

\item{default.alpha}{a value between 0 and 1 indicating alpha to be used if no alpha aesthetics is used.}

\item{range.alpha}{a numeric vector of length 2 with lower and upper limit of alpha range.}

\item{viridis.option}{a character vector indicating Viridis color scale to be used,
options include \emph{viridis} (default), \emph{magma}, \emph{plasma}, \emph{inferno}, \emph{cividis}.}

\item{viridis.direction}{a direction of the Viridis color palette.}

\item{viridis.range}{a numeric vector of length 2 with lower and upper limit of the Viridis color palette.}

\item{label.alpha}{a value between 0 and 1 indicating alpha of the labels.}

\item{segment.alpha}{a value between 0 and 1 indicating alpha of the labels repels.}

\item{repel.xynudges}{a numeric vector of length 2 with horizontal and vertical adjustments to nudge the position of labels.}

\item{repel.hvjust}{a numeric vector of length 2 with horizontal and vertical justification of labels.}

\item{repel.direction}{a character indicating direction in which to adjust position of labels, options are \emph{both} (default), \emph{x} and \emph{y}.}

\item{base.fontsize}{an integer indicating base font size.}

\item{size.guide.fontsize}{an integer indicating size guide font size.}

\item{label.fontsize}{an integer indicating labels font size.}

\item{title}{a text for the title.}

\item{subtitle}{a text for the subtitle.}

\item{xlabel}{a text for the x-axis label.}

\item{ylabel}{a text for the y-axis label.}

\item{xlimits}{a numeric vector of length 2 indicating lower and upper x-axis limits}

\item{ylimits}{a numeric vector of length 2 indicating lower and upper y-axis limits}

\item{void}{a boolean indicating whether theme void should be used.}
}
\value{
A gsoap plot, where individual gene sets are represented by circles,
whose size reflects the number of query gene members of the given set,
and whose mutual proximities reflect the number common gene members.
}
\description{
Some description
}
\details{
More description
}
\examples{
data(pxgenes)

l = gsoap_layout(pxgenes, 'Members', 'p.value')
gsoap_plot(l, as.color = 'cluster', as.alpha = 'closeness')

}
\author{
Tomas Tokar <tomastokar@gmail.com>
}
